depth working ??

+++ src/util/lerp.h
@@ -1,7 +1,8 @@
 #include <vector>
 #include <glm/glm.hpp>
+#include <iostream>
 
-std::vector<float> interpolateSingleFloats(float, float, float);
+std::vector<float> interpolateSingleFloats(float, float, int);
 
 
 std::vector<glm::vec3> interpolateThreeElementValues(glm::vec3, glm::vec3, float);
